blog/requirements.txt
asttokens==2.4.0
charset-normalizer==3.3.2
Django==4.2.1
django-debug-toolbar==4.2.0
django-extensions==3.2.3
django-filter==24.2
django-simple-captcha==0.5.18
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.0
drf-spectacular==0.27.2
flower==2.0.0
gunicorn==21.2.0
hiredis==2.3.2
humanize==4.9.0
ipython==8.16.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
Pillow==10.0.1
psycopg2-binary
PyJWT==2.8.0
redis==5.0.1
setuptools==69.2.0
social-auth-app-django==5.3.0
social-auth-core==4.4.2
typing_extensions==4.11.0
Unidecode==1.3.8
celery[redis]==5.3.5